requirements.txt
Click>=6.7
PyYAML>=3.10
google-cloud-storage>=1.7.0
cassandra-driver>=3.14.0
paramiko>=2.4.1
psutil>=5.4.7
ffwd>=0.0.2
apache-libcloud>=2.4.0
python-dateutil>=2.8.0
lockfile>=0.12.2